checkpoint: renames "Server/ServerSpecificationError.ts" to "Server/Error/Specification.ts"
