+++ src/features/TextToImage/Server/Error/Specification.ts
